+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [2.1.0] - 2023-09-12
+- The NewEnvi() function returns the envi struct itself instead of an interface
+- Added functionality for easy loading yaml, json and plain content in files from file paths that are set in environment variables
+
 ## [2.0.0] - 2021-02-11
 
 This is more or less a complete rewrite of the envi package.